refactor(core): make permission/mod.rs re-export-only
Every value-object family keeps its type in a leaf file and uses mod.rs
purely for re-exports; the permission family was the exception, with the
455-line `Permission` enum, its impl, the catalog helpers and their tests
living directly in mod.rs.

Move all of it into `permission/catalog.rs` (the enum + impl +
`PERMISSION_CATALOG` / `is_known_permission` / `permission_resource_type`
+ the catalog tests) and reduce mod.rs to re-exports. Public paths
(`permission::Permission`, `permission::ResourceRef`,
`permission::is_known_permission`, ...) are unchanged, so no call site
moves. No behaviour change.
